Add Volumetric Rendering to Archive, Mandelbulb Distance Estimator, Create "Other Experiments" Folder, and New Default Scene
The Volumetric Rendering has been added to the archive folder since it is a milestone, and the Mandelbulb Distance estimator is a (Currently Unused) Distance estimator. The new "Other Experiments" Folder is for other related experiments with C and Path-Tracing like programs. Last in these additions is a new default scene which includes a third material on a neat looking object.