Volumetric Folder, Noise, and Simple-Planet scene.
This isn't a main part of the path-tracer, it will be in the future but for now we'll store it in a special folder. Also, there is now FBM noise that takes a vec3 and outputs a float. Last, there is a new Simple scene with a Planet that makes use of this.